requirements.txt
python-docx
pdfminer.six
docx2txt
pandas
python-pptx